+++ dev_handler/devices.c
