+++ dev_handler/message.c
